bt_follow_me/visualizer/pyproject.toml
[project]
name = "bt-follow-me-visualizer"
version = "0.1.0"
description = "Visualizer for Bluetooth AoA data from the FollowMe project"
requires-python = ">=3.8"
dependencies = [
    "paho-mqtt>=1.6.1",
    "matplotlib>=3.7.1",
    "numpy>=1.24.3"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."] 